homework2/requirements.txt
asgiref==3.11.1
behave==1.3.3
colorama==0.4.6
cucumber-expressions==19.0.0
cucumber-tag-expressions==9.1.0
Django==6.0.2
djangorestframework==3.16.1
parse==1.21.1
parse_type==0.6.6
six==1.17.0
sqlparse==0.5.5
